refactor(ingestion-consumer): move the commit sentinel into its own module
order_sentinel.rs keeps the per-key send-order sentinel and the rdkafka
context; the commit sentinel and its tests move to commit_sentinel.rs,
next to the commit manager it observes. No behaviour changes.
