+++ rust/ingestion-consumer/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod aperture;
 pub mod batcher;
 pub mod commit_manager;
+pub mod commit_sentinel;
 pub mod config;
 pub mod consumer;
 pub mod debug_recorder;
